+++ nudebomb/__init__.py
@@ -1,1 +1,8 @@
 """Nudebomb."""
+
+from os import environ
+
+if environ.get("PYTHONDEVMODE"):
+    from icecream import install  # pyright: ignore[reportPrivateImportUsage]
+
+    install()